Right we have basically followed the code:
What we have is data files from Rory that have a geoid variable that should be made from state, county, and tract but in reality doesnt.

And we have shp files we downloaded ourselves.

So what we have to do is stitch all the shp files into one file of tracts for US and join that to data from Rory on the largest MSA.

What we found is that the shp files together have duplicated state, county and tract info but small duplicates. Furthermore there are issues regarding why so data points are not matched with their shp file equivalents.

Nonetheless we have done joins for 2000 and 2009 so we can practise on code in the exampple folder